requirements.txt
flask==2.3.2
langchain==0.1.5
langchain-community==0.0.10
langchain-google-genai==0.0.5
langchain-anthropic==0.1.1
openai==1.6.1
google-search-results==2.4.2
python-dotenv==1.0.0
google-generativeai==0.3.1
anthropic==0.7.0
